// src/enums/EnumSCPICommands.ts
export enum EnumSCPICommands {
  RESET = '*RST',
  GET_ID = '*IDN?',
  SELF_TEST = '*TST?',
  REMOTE = 'SYSTem:REMote',
  LOCAL = 'SYSTem:LOCal',
  SET_OUTPUT = 'OUTPut',
  SET_OUTPUT_ON = 'OUTPut ON',
  SET_OUTPUT_OFF = 'OUTPut OFF',
  GET_OUTPUT = 'OUTPut?',
  MEASURE_VOLTAGE = 'MEASure:VOLTage?',
  MEASURE_CURRENT = 'MEASure:CURRent?',
  MEASURE_POWER = 'MEASure:POWer?',
  MEASURE_ALL = 'MEASure:ALL?',
  MEASURE_ALL_INFO = 'MEASure:ALL:INFO?',
  GET_CURRENT = 'CURRent?',
  SET_CURRENT = 'CURRent',
  GET_CURRENT_LIMIT = 'CURRent:LIMit?',
  SET_CURRENT_LIMIT = 'CURRent:LIMit',
  // SET_CURRENT_LIMIT = 'CURR:LIM',
  GET_VOLTAGE = 'VOLTage?',
  SET_VOLTAGE = 'VOLTage',
  GET_VOLTAGE_LIMIT = 'VOLTage:LIMit?',
  SET_VOLTAGE_LIMIT = 'VOLTage:LIMit',
}